requirement.txt
Django==3.0.3
django-allauth==0.41.0
djangorestframework==3.11.0
Markdown==3.2.2
Pillow==7.0.0
python==3.7